Restructuring template schamas directories
