+++ spec/components/schemas/templates/components/buttons/base.ts
@@ -1,5 +1,5 @@
 import { SchemaObject } from 'openapi3-ts';
-import { createComponentRef } from '../../../../utils/ref';
+import { createComponentRef } from '../../../../../../utils/ref';
 
 const contentBase: SchemaObject = {
   type: 'object',
