+++ spec/components/schemas/templates/components/buttons/button-item-base.ts
@@ -1,0 +1,26 @@
+import { SchemaObject } from 'openapi3-ts';
+import { createComponentRef } from '../../../../../../utils/ref';
+
+const buttonItemBase: SchemaObject = {
+  type: 'object',
+  properties: {
+    type: {
+      title: 'Button type',
+      enum: [
+        'URL', 'PHONE_NUMBER', 'QUICK_REPLY',
+      ],
+      type: 'string',
+    },
+    text: {
+      title: 'Button text',
+      type: 'string',
+    },
+  },
+  required: [
+    'type',
+    'text',
+  ],
+};
+
+export const ref = createComponentRef(__filename);
+export default buttonItemBase;
