+++ spec/components/schemas/templates/components/buttons/button-item-url.ts
@@ -1,20 +1,15 @@
 // tslint:disable:max-line-length
 import { SchemaObject } from 'openapi3-ts';
-import { createComponentRef } from '../../../../utils/ref';
-import { ref as baseRef } from './base';
+import { createComponentRef } from '../../../../../../utils/ref';
+import { ref as baseRef } from './button-item-base';
 
-const buttons: SchemaObject = {
+const urlButton: SchemaObject = {
   type: 'object',
   allOf: [{
     $ref: baseRef,
   }, {
     type: 'object',
     properties: {
-      text: {
-        title: 'Button text',
-        description: 'Text that will fill the button',
-        type: 'string',
-      },
       url: {
         title: 'URL of button',
         description: 'URL to be sent in the button. It can be dynamic',
@@ -22,6 +17,7 @@ const buttons: SchemaObject = {
       },
     },
     required: [
+      'type',
       'text',
       'url',
     ],
@@ -29,4 +25,4 @@ const buttons: SchemaObject = {
 };
 
 export const ref = createComponentRef(__filename);
-export default buttons;
+export default urlButton;
