+++ spec/components/schemas/templates/components/header/index.ts
@@ -1,0 +1,20 @@
+import { SchemaObject } from 'openapi3-ts';
+import { createComponentRef } from '../../../../../../utils/ref';
+
+export const header: SchemaObject = {
+  title: 'Header',
+  description: 'Template header',
+  type: 'object',
+  required: ['type'],
+  properties: {
+    type: {
+      title: 'Header type',
+      description: 'The header can have an image, document or text.',
+      type: 'string',
+      enum: ['MEDIA_DOCUMENT', 'MEDIA_IMAGE', 'TEXT_FIXED', 'TEXT_TEMPLATE'],
+    },
+  },
+};
+
+export const ref = createComponentRef(__filename);
+export default header;
